+++ sentinelsat/scripts/cli.py
@@ -10,7 +10,6 @@ import requests.utils
 from sentinelsat import __version__ as sentinelsat_version
 
 from sentinelsat.sentinel import (
-    SentinelAPI,
     geojson_to_wkt,
     read_geojson,
     placename_to_wkt,
@@ -18,7 +17,7 @@ from sentinelsat.sentinel import (
 )
 
 from sentinelsat.exceptions import InvalidKeyError
-from sentinelsat.advanced import AdvancedSentinelAPI, make_path_filter
+from sentinelsat.products import SentinelProductsAPI, make_path_filter
 
 
 json_parse_exception = json.decoder.JSONDecodeError
@@ -241,7 +240,7 @@ def cli(
             "for environment variables and .netrc support."
         )
 
-    api = AdvancedSentinelAPI(user, password, url)
+    api = SentinelProductsAPI(user, password, url)
 
     if info:
         ctx = click.get_current_context()