+++ tests/test_download.py
@@ -331,13 +331,13 @@ def test_get_stream(api, tmpdir, smallest_online_products):
 
 @pytest.mark.vcr
 @pytest.mark.scihub
-def test_download_advanced(advanced_api, tmpdir, smallest_online_products):
+def test_download_product_nodes(products_api, tmpdir, smallest_online_products):
     uuid = smallest_online_products[0]["id"]
     product_dir = smallest_online_products[0]["title"] + ".SAFE"
     expected_path = tmpdir.join(product_dir)
 
     nodefilter = make_path_filter("*preview/*.kml")
-    product_info = advanced_api.download(uuid, str(tmpdir), checksum=True, nodefilter=nodefilter)
+    product_info = products_api.download(uuid, str(tmpdir), checksum=True, nodefilter=nodefilter)
 
     assert os.path.normpath(product_info["node_path"]) == product_dir
     assert expected_path.exists()